+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -248,7 +248,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01ke92zwsqyw2wftn85z2b17vq"),
+  _id: z.string().default("conv_01ke9dbnv2vqwr3ew3tgvg64ck"),
   entityId: z.string(),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
@@ -282,7 +282,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01ke92zwsqyw2wftn85z2b17vq"),
+  id: z.string().default("conv_01ke9dbnv2vqwr3ew3tgvg64ck"),
   entityId: z.string(),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),