+++ packages/orq-rc/RELEASES.md
@@ -5709,3 +5709,13 @@ Based on:
 - [typescript v4.2.0-rc.10] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.10] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.10 - packages/orq-rc
+
+## 2026-01-06 10:26:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.681.0 (2.789.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.11] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.11 - packages/orq-rc